+++ sportsdataverse/__init__.py
@@ -54,6 +54,7 @@ from sportsdataverse.mlb import *
 from sportsdataverse.nba import *
 from sportsdataverse.nfl import *
 from sportsdataverse.nhl import *
+from sportsdataverse.pwhl import *
 from sportsdataverse.wbb import *
 from sportsdataverse.wnba import *